+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-delivery-order-impl/src/main/java/com/liferay/headless/commerce/delivery/order/internal/dto/v1_0/converter/PlacedOrderCommentDTOConverter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.headless.commerce.delivery.order.internal.dto.v1_0;
+package com.liferay.headless.commerce.delivery.order.internal.dto.v1_0.converter;
 
 import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.model.CommerceOrderNote;
