+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-delivery-order-impl/src/main/java/com/liferay/headless/commerce/delivery/order/internal/resource/v1_0/PlacedOrderCommentResourceImpl.java
@@ -108,7 +108,7 @@ public class PlacedOrderCommentResourceImpl
 	private DTOConverterRegistry _dtoConverterRegistry;
 
 	@Reference(
-		target = "(component.name=com.liferay.headless.commerce.delivery.order.internal.dto.v1_0.PlacedOrderCommentDTOConverter)"
+		target = "(component.name=com.liferay.headless.commerce.delivery.order.internal.dto.v1_0.converter.PlacedOrderCommentDTOConverter)"
 	)
 	private DTOConverter<CommerceOrderNote, PlacedOrderComment>
 		_placedOrderCommentDTOConverter;
